+++ src/app/icons/v2/index.tsx
@@ -18,6 +18,7 @@ import { ReactComponent as ArrowRightUp } from './arrow-right-up.svg';
 import { ReactComponent as ArrowRight } from './arrow-right.svg';
 import { ReactComponent as ArrowUpDown } from './arrow-up-down.svg';
 import { ReactComponent as ArrowUp } from './arrow-up.svg';
+import { ReactComponent as BackArrow } from './back-arrow.svg';
 import { ReactComponent as Backspace } from './backspace.svg';
 import { ReactComponent as Bin } from './bin.svg';
 import { ReactComponent as Calendar } from './calendar.svg';
@@ -109,6 +110,7 @@ export enum IconName {
   ArrowRight = 'arrow-right',
   ArrowUpDown = 'arrow-up-down',
   ArrowUp = 'arrow-up',
+  BackArrow = 'back-arrow',
   Backspace = 'backspace',
   Calendar = 'calendar',
   Bin = 'bin',
@@ -220,6 +222,8 @@ const IconSwitch = (props: IconProps) => {
       return <ArrowUpDown {...props} />;
     case IconName.ArrowUp:
       return <ArrowUp {...props} />;
+    case IconName.BackArrow:
+      return <BackArrow {...props} />;
     case IconName.Backspace:
       return <Backspace {...props} />;
     case IconName.Calendar: